[Chore] fixes the incorrect information my previous commit introduced.
The only reason a user's nick shouldn't appear in a leaderboard that
was generated with the shift key is if no tournaments have finished
that the user was entered in.  IOW, if someone creates a new nick,
enters a tournament and loses, that player won't show up with a
negative number in the leaderboard even if the user holds the shift
key until that tournament has finished.