DO NOT MERGE: Partial Solution that doesn't work for HTML
This was part of my exploration of the technique described on
tex.stackexchange.com at
https://tex.stackexchange.com/questions/497403/how-to-use-noto-color-emoji-with-lualatex

It produces OK results in the PDF, but the HTML that is created is
fried.  Perhaps there's a pre-processor that could be used to switch
the TeX code in when doing a TeX build and the original code in when
doing an HTML build, but it would still have the draw-back that the
technique doesn't work for inline emojis; instead the entire block has
to be a TeX block which means that any other style changes have to be
done in TeX rather than markdown.

I do not plan on merging this, but there's a tiny chance that the
technique could be useful for the revamped BARGE Rule Book, although
if it's not done in mdbook, having header-includes in toml rather than
yaml is sub-optimal.