DO NOT MERGE: squash -- I copied some text from an email so I wouldn't lose it
It's not uncommon for me to write something very quickly in email that belongs
in the Players' Manual, only for me to forget to actually include it (sometimes
even forgetting that I've even written it).  I think I've described the
impossible to rig deck more than once, but since I just described it again,
I've put that text in here as a placeholder / reminder.